--- serie2/src/serie2/Error.java
@@ -1,8 +1,0 @@
-package serie2;
-public class Error {
-
-    public static void main(String[] args) {
-      System.out.println("Correct Me");
-    }
-  
-}
