+++ week2/src/serie2/Degre3.java
