+++ week2/src/serie2/Fondue.java
