+++ week2/src/serie2/Moyenne1.java
