+++ week2/src/serie2/Moyenne2.java
