+++ week2/src/serie2/Permutation.java
